requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pytest==7.4.3
pydantic==2.9.2
python-multipart==0.0.6
aiofiles==23.2.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
httpx==0.27.0



